==> 01.Wrangling.R <==
library(tidyverse)
library(suncalc)
library(lubridate)
library(data.table)
library(adehabitatLT)
library(meanShiftR) #for density cluster analysis
library(sf)

options(scipen = 999)

whemi <- map_data("world", region=c("Canada", 
                                    "USA", 
                                    "Mexico",
                                    "Guatemala", 
                                    "Belize", 
                                    "El Salvador",
                                    "Honduras", 
                                    "Nicaragua", 
                                    "Costa Rica",
                                    "Panama", 
                                    "Jamaica", 
                                    "Cuba", 
                                    "The Bahamas",
                                    "Haiti", 
                                    "Dominican Republic", 
                                    "Antigua and Barbuda",
                                    "Dominica", 
                                    "Saint Lucia", 
                                    "Saint Vincent and the Grenadines", 
                                    "Barbados",
                                    "Grenada",
                                    "Trinidad and Tobago",
                                    "Colombia",
                                    "Venezuela",
                                    "Guyana",
                                    "Suriname",
                                    "Ecuador",
                                    "Peru",
                                    "Brazil",
                                    "Bolivia",
                                    "Paraguay",
                                    "Chile",
                                    "Argentina",
                                    "Uruguay")) %>% 
  dplyr::filter(!group%in%c(258:264))

#1. Read in data----
dat <- read.csv("/Users/ellyknight/Documents/UoA/Projects/Projects/MCP2/Analysis/Data/CONIMCP_CleanDataAll.csv") %>% 
  mutate(Date = as.Date(str_sub(DateTime, 1, 10)),
         DateTime = ymd_hms(paste0(Date, Time))) %>% 
  dplyr::select(PinpointID, Population, Mass, Wing, Sex, Type, DateTime, Date, Time, Year, Lat, Long, BandDist, WintDist, GCD, Season, Season2, Winter) %>% 
  unique()
  
#2. Calculate sun times----
dat.sun <- getSunlightTimes(data=dat%>% 
                               rename(lat = Lat, lon = Long, date=Date) %>% 
                               dplyr::select(date, lat, lon)) %>% 
  dplyr::select(sunrise, sunset) %>% 
  cbind(dat) %>% 
  mutate(sun = ifelse(DateTime > sunrise & DateTime < sunset, 1, 0))

#3. Calculate NSD----
traj <- as.ltraj(xy=dat.sun[,c("Long", "Lat")],
                 id=dat.sun$PinpointID,
                 date=dat.sun$DateTime,
                 proj4string = CRS("+proj=longlat +datum=WGS84")) 

dat.traj <- rbindlist(traj) %>% 
  dplyr::select(dist, R2n, abs.angle, rel.angle) %>% 
  cbind(dat.sun) %>% 
  mutate(doy=yday(DateTime)) %>% 
  group_by(PinpointID) %>% 
  mutate(R2n2 = max(R2n) - R2n) %>% 
  ungroup()

#4. Prepare for clustering----
dat.coord <- dat.traj %>% 
  st_as_sf(crs=4326, coords=c("Long", "Lat")) %>% 
  st_transform(crs=3857) %>% 
  st_coordinates() %>% 
  cbind(dat.traj)

ids <- dat.coord %>% 
  dplyr::select(PinpointID) %>% 
  unique()

#5. Mean shift classification----
dat.shift <- data.frame()
for(i in 1:nrow(ids)){
  
  dat.i <- dat.coord %>% 
    dplyr::filter(PinpointID==ids$PinpointID[i])
  
  mat1 <- matrix(dat.i$X)
  mat2 <- matrix(dat.i$Y)
  mat <- cbind(mat1, mat2)
  
  shift <- meanShift(mat,
                         algorithm="KDTREE",
                         bandwidth=c(1,1))
  
  dat.shift <- dat.i %>% 
    mutate(cluster = shift[[1]]) %>% 
    rbind(dat.shift)
}

#6. Remove points that aren't in a cluster----
dat.clust <- dat.shift %>% 
  group_by(PinpointID, cluster) %>% 
  mutate(count=n()) %>% 
  ungroup() %>% 
  mutate(stopover = ifelse(count > 3, 1, 0)) %>% 
  dplyr::filter(stopover==1)

#7. Clean wintering data----
dat.wint <- dat.clust %>% 
  dplyr::filter(Lat < 25) %>% 
  group_by(PinpointID, cluster) %>% 
  mutate(minDate = min(DateTime),
         maxDate = max(DateTime),
         duration = maxDate - minDate) %>% 
  ungroup() %>% 
  dplyr::filter(duration > 21) %>% 
  mutate(Season="Winter") %>% 
  dplyr::select(-Winter)

#Visualize
plot.wint <- ggplot() +
  geom_point(data=dat.wint, aes(x=Long, y=Lat, colour=doy), size=3, alpha=0.7) +
  labs(x = "", y = "") +
  theme_bw() +
  scale_colour_viridis_c() +
  facet_wrap(~PinpointID, ncol=10, scales="free")

ggsave(plot.wint, file="Figures/MeanShift_wint.jpeg", width=20, height=10, unit="in")

#individuals with 2 wintering grounds
ids.wint <- dat.wint %>% 
  dplyr::select(PinpointID, Season, cluster) %>% 
  unique() %>% 
  group_by(PinpointID, Season) %>% 
  summarize(n=n()) %>% 
  dplyr::filter(n==2) %>% 
  left_join(dat.wint %>% 
              dplyr::select(PinpointID, Season, cluster, minDate) %>% 
              unique()) %>% 
  mutate(Winter = ifelse(minDate==min(minDate), 1, 2)) %>% 
  ungroup()

dat.wint2 <- dat.wint %>% 
  left_join(ids.wint) %>% 
  mutate(Winter = ifelse(is.na(Winter), 1, Winter)) %>% 
  dplyr::select(PinpointID, Population, Mass, Wing, Sex, Type, DateTime, Date, doy, Time, sun, Year, Lat, Long, BandDist, WintDist, GCD, Season, Winter, dist, R2n, abs.angle, rel.angle, cluster, count)

#8. Clean breeding data----
dat.ids <- dat.wint %>% 
  dplyr::select(PinpointID) %>% 
  unique()

dat.breed <- dat.traj %>% 
  dplyr::filter(PinpointID %in% dat.ids$PinpointID) %>% 
  left_join(dat.clust) %>% 
  dplyr::filter(Lat > 25,
                stopover==1 | Type=="Band") %>% 
  group_by(PinpointID) %>% 
  mutate(maxpt = ifelse(Lat==max(Lat), cluster, 0),
         maxclust = max(maxpt),
         useclust = ifelse(cluster==maxclust, 1, 0)) %>% 
  mutate(TypeNo = ifelse(Type=="Band", 0, 1),
         TypeSum = sum(TypeNo),
         BandUse = case_when(TypeSum==0 ~ 1,
                             is.na(maxclust) & Type=="Band" ~ 1,
                             PinpointID %in% c(480, 489) ~1)) %>% 
  ungroup() %>% 
  mutate(Season="Breed",
         Winter=0)  %>% 
  dplyr::filter((useclust==1 & Type != "Band") | BandUse==1) %>% 
  dplyr::select(PinpointID, Population, Mass, Wing, Sex, Type, DateTime, Date, doy, Time, sun, Year, Lat, Long, BandDist, WintDist, GCD, Season, Winter, dist, R2n, abs.angle, rel.angle, cluster, count)

#Visualize
plot.breed <- ggplot() +
  geom_point(data=dat.breed, aes(x=Long, y=Lat, colour=factor(Type)), size=3, alpha=0.7) +
  labs(x = "", y = "") +
  theme_bw() +
  scale_colour_viridis_d() +
  facet_wrap(~PinpointID, ncol=10, scales="free")

ggsave(plot.breed, file="Figures/MeanShift_Breed.jpeg", width=20, height=10, unit="in")

#9. Take out some points manually----
dat.season <- rbind(dat.wint2, dat.breed) %>% 
  dplyr::filter(!(PinpointID==489 & Lat > -8.5 & Type!="Band"),
                !(PinpointID==480 & Long < -67 & Type!="Band"),
                !(PinpointID==443 & Long < -28.89))

#10. Final product----
table(dat.season$Season, dat.season$PinpointID)
write.csv(dat.season, "CONIMCP_CleanDataAll_Habitat.csv", row.names = FALSE)

dat.roost <- dat.season %>% 
  dplyr::filter(sun==1 | Type=="Band")

table(dat.roost$Season, dat.roost$PinpointID)

write.csv(dat.roost, "CONIMCP_CleanDataAll_Habitat_Roosting.csv", row.names = FALSE)

#Visualize
plot.shift <- ggplot() +
  geom_polygon(data=whemi, aes(x=long, y=lat, group=group), colour = "gray85", fill = "gray75", size=0.3) +
  geom_point(data=dat.roost, aes(x=Long, y=Lat, colour=factor(Season)), size=3, alpha=0.7) +
  labs(x = "", y = "") +
  xlim(c(-170, -30)) +
  theme_bw() +
  scale_colour_viridis_d() +
  facet_wrap(~PinpointID, ncol=10)

ggsave(plot.shift, file="Figures/MeanShift_Map.jpeg", width=20, height=10, unit="in")

#11. Final numebrs----
length(unique(dat.roost$PinpointID))
length(unique(dat.roost$Population))
nrow(dat.roost)
